social-cinema-api/requirements.txt
alembic==1.3.1
astroid==1.6.6
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==19.7.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pipenv==2018.11.26
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.0
PyJWT==1.7.1
pylint==1.8.2
pymongo==3.3.1
pytest==3.0.6
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.11.2
